+++ pkg/idp/provider/azure/azure.go
@@ -17,16 +17,24 @@ import (
 	msgraphsdk "github.com/microsoftgraph/msgraph-sdk-go"
 	"github.com/microsoftgraph/msgraph-sdk-go/applications/item/removepassword"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
+	"github.com/skratchdot/open-golang/open"
 	"gopkg.in/yaml.v2"
 )
 
 type Azure struct {
-	Name     string
-	Client   *msgraphsdk.GraphServiceClient
-	Log      *logr.Logger
-	Owner    string
-	TenantID string
-	Type     string
+	Name         string
+	Client       *msgraphsdk.GraphServiceClient
+	Log          *logr.Logger
+	Owner        string
+	TenantID     string
+	Type         string
+	clientSecret string
+}
+
+type Config struct {
+	TenantID     string
+	ClientID     string
+	ClientSecret string
 }
 
 type config struct {
@@ -45,7 +53,8 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Azure, error) {
 
 	var client *msgraphsdk.GraphServiceClient
 	{
-		cred, err := azidentity.NewClientSecretCredential(c.tenantID, c.clientID, c.clientSecret, nil)
+
+		cred, err := azidentity.NewClientSecretCredential(c.TenantID, c.ClientID, c.ClientSecret, nil)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -63,48 +72,54 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Azure, error) {
 		}
 	}
 	return &Azure{
-		Name:     key.GetProviderName(p.Owner, p.Name),
-		Log:      log,
-		Type:     ProviderConnectorType,
-		Client:   client,
-		Owner:    p.Owner,
-		TenantID: c.tenantID,
+
+		Name:         key.GetProviderName(p.Owner, p.Name),
+		Log:          log,
+		Type:         ProviderConnectorType,
+		Client:       client,
+		Owner:        p.Owner,
+		TenantID:     c.TenantID,
+		clientSecret: c.ClientSecret,
 	}, nil
 }
 
-func newAzureConfig(p provider.ProviderCredential, log *logr.Logger) (config, error) {
+func newAzureConfig(p provider.ProviderCredential, log *logr.Logger) (Config, error) {
 	if log == nil {
-		return config{}, microerror.Maskf(invalidConfigError, "Logger must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Logger must not be empty.")
 	}
 	if p.Name == "" {
-		return config{}, microerror.Maskf(invalidConfigError, "Credential name must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Credential name must not be empty.")
 	}
 	if p.Owner == "" {
-		return config{}, microerror.Maskf(invalidConfigError, "Credential owner must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Credential owner must not be empty.")
 	}
 
 	var tenantID, clientID, clientSecret string
 	{
 		if tenantID = p.Credentials[TenantIDKey]; tenantID == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", TenantIDKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", TenantIDKey)
 		}
 		if clientID = p.Credentials[ClientIDKey]; clientID == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientIDKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientIDKey)
 		}
 		if clientSecret = p.Credentials[ClientSecretKey]; clientSecret == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientSecretKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientSecretKey)
 		}
 	}
 
-	return config{
-		tenantID:     tenantID,
-		clientID:     clientID,
-		clientSecret: clientSecret,
+	return Config{
+		TenantID:     tenantID,
+		ClientID:     clientID,
+		ClientSecret: clientSecret,
 	}, nil
 }
 
 func (a *Azure) GetName() string {
 	return a.Name
+}
+
+func (a *Azure) GetProviderName() string {
+	return ProviderName
 }
 
 func (a *Azure) GetType() string {
@@ -129,7 +144,7 @@ func (a *Azure) CreateOrUpdateApp(config provider.AppConfig, ctx context.Context
 	}
 
 	//Create or update secret
-	secret, err := a.createOrUpdateSecret(id, config, ctx, oldSecret)
+	secret, err := a.CreateOrUpdateSecret(id, config, ctx, oldSecret, false)
 	if err != nil {
 		return provider.ProviderApp{}, microerror.Mask(err)
 	}
@@ -194,7 +209,7 @@ func (a *Azure) createOrUpdateApplication(config provider.AppConfig, ctx context
 	return *id, nil
 }
 
-func (a *Azure) createOrUpdateSecret(id string, config provider.AppConfig, ctx context.Context, oldSecret string) (provider.ProviderSecret, error) {
+func (a *Azure) CreateOrUpdateSecret(id string, config provider.AppConfig, ctx context.Context, oldSecret string, skipDelete bool) (provider.ProviderSecret, error) {
 
 	app, err := a.Client.ApplicationsById(id).Get(ctx, nil)
 	if err != nil {
@@ -217,26 +232,37 @@ func (a *Azure) createOrUpdateSecret(id string, config provider.AppConfig, ctx c
 
 	// We delete the secret in case it exists and is expired or in case we do not have the key anymore
 	if !needsCreation && (!keyPresent || secretExpired(secret) || secretChanged(secret, oldSecret)) {
-		requestBody := removepassword.NewRemovePasswordPostRequestBody()
-		requestBody.SetKeyId(secret.GetKeyId())
-
-		err = a.Client.ApplicationsById(id).RemovePassword().Post(context.Background(), requestBody, nil)
-		if err != nil {
-			return provider.ProviderSecret{}, microerror.Maskf(requestFailedError, PrintOdataError(err))
+		if skipDelete {
+			a.Log.Info(fmt.Sprintf("Skipped deletion of secret %v of app %s in microsoft ad tenant %s", secret.GetKeyId(), id, a.TenantID))
+		} else {
+			if err = a.DeleteSecret(ctx, secret.GetKeyId(), id); err != nil {
+				return provider.ProviderSecret{}, microerror.Mask(err)
+			}
+			a.Log.Info(fmt.Sprintf("Removed secret %v of %s app %s for %s in microsoft ad tenant %s", secret.GetKeyId(), a.Type, config.Name, a.Owner, a.TenantID))
 		}
-		a.Log.Info(fmt.Sprintf("Removed secret %v of %s app %s for %s in microsoft ad tenant %s", secret.GetKeyId(), a.Type, config.Name, a.Owner, a.TenantID))
 		needsCreation = true
 	}
 
 	// Create secret if it does not exist
 	if needsCreation {
-		secret, err = a.Client.ApplicationsById(id).AddPassword().Post(ctx, GetSecretCreateRequestBody(config.Name, key.SecretValidityMonths), nil)
+		secret, err = a.Client.ApplicationsById(id).AddPassword().Post(ctx, GetSecretCreateRequestBody(config), nil)
 		if err != nil {
 			return provider.ProviderSecret{}, microerror.Maskf(requestFailedError, PrintOdataError(err))
 		}
 		a.Log.Info(fmt.Sprintf("Created secret %v of %s app %s for %s in microsoft ad tenant %s", secret.GetKeyId(), a.Type, config.Name, a.Owner, a.TenantID))
 	}
 	return getAzureSecret(secret, app, oldSecret)
+}
+
+func (a *Azure) DeleteSecret(ctx context.Context, secretID *string, appID string) error {
+	requestBody := removepassword.NewRemovePasswordPostRequestBody()
+	requestBody.SetKeyId(secretID)
+
+	err := a.Client.ApplicationsById(appID).RemovePassword().Post(ctx, requestBody, nil)
+	if err != nil {
+		return microerror.Maskf(requestFailedError, PrintOdataError(err))
+	}
+	return nil
 }
 
 func (a *Azure) DeleteApp(name string, ctx context.Context) error {
@@ -317,4 +343,83 @@ func (a *Azure) computeAppUpdatePatch(config provider.AppConfig, app models.Appl
 	}
 
 	return appNeedsUpdate, appPatch
+}
+
+// TODO
+// improve output
+// include new service principal creation
+func (a *Azure) GetCredentialsForAuthenticatedApp(config provider.AppConfig) (string, error) {
+	ctx := context.Background()
+	app, err := a.GetApp(config.Name)
+	if err != nil {
+		if !IsNotFound(err) {
+			return "", microerror.Mask(err)
+		}
+		// Create app if it does not exist
+		app = models.NewApplication()
+		app.SetDisplayName(&config.Name)
+
+		// Set permissions from parent app
+		parentApp, err := a.GetApp(DexOperatorName)
+		if err != nil {
+			return "", microerror.Mask(err)
+		}
+		if needsUpdate, patch := computePermissionsUpdatePatch(app, parentApp); needsUpdate {
+			app.SetRequiredResourceAccess(patch)
+		}
+		app, err = a.Client.Applications().Post(ctx, app, nil)
+		if err != nil {
+			return "", microerror.Maskf(requestFailedError, PrintOdataError(err))
+		}
+		a.Log.Info(fmt.Sprintf("Created %s app %s for %s in microsoft ad tenant %s", a.Type, config.Name, a.Owner, a.TenantID))
+		if app.GetAppId() == nil {
+			return "", microerror.Maskf(notFoundError, "Could not find client ID of app %s.", config.Name)
+		}
+		consentURL := getAdminConsentUrl(a.TenantID, *app.GetAppId())
+		a.Log.Info(fmt.Sprintf("Admin consent is needed. Please grant under the following URL: %s", consentURL))
+		a.Log.Info("Please be aware that it can take a while for the app to become available. Wait a moment before logging in and granting consent.")
+		err = open.Start(consentURL)
+		if err != nil {
+			return "", microerror.Mask(err)
+		}
+	}
+
+	id := app.GetId()
+	if id == nil {
+		return "", microerror.Maskf(notFoundError, "Could not find ID of app %s.", config.Name)
+	}
+	secret, err := a.CreateOrUpdateSecret(*id, config, ctx, "", true)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	return fmt.Sprintf(`client-id: %s
+client-secret: %s
+tenant-id: %s`, secret.ClientId, secret.ClientSecret, a.TenantID), nil
+}
+
+func (a *Azure) CleanCredentialsForAuthenticatedApp(config provider.AppConfig) error {
+	for _, name := range []string{config.Name, DexOperatorName} {
+		app, err := a.GetApp(name)
+		if err != nil {
+			if !IsNotFound(err) {
+				return microerror.Mask(err)
+			}
+			continue
+		}
+		id := app.GetId()
+		if id == nil {
+			return microerror.Maskf(notFoundError, "Could not find ID of app %s.", name)
+		}
+		for _, c := range app.GetPasswordCredentials() {
+			if credentialName := c.GetDisplayName(); credentialName != nil {
+				if *credentialName == config.Name && secretChanged(c, a.clientSecret) {
+					if err = a.DeleteSecret(context.Background(), c.GetKeyId(), *id); err != nil {
+						return microerror.Mask(err)
+					}
+					a.Log.Info(fmt.Sprintf("Removed secret %v of %s app %s for %s in microsoft ad tenant %s", c.GetKeyId(), a.Type, config.Name, a.Owner, a.TenantID))
+				}
+			}
+		}
+	}
+	return nil
 }