+++ pkg/idp/provider/azure/azure_test.go
@@ -2,7 +2,6 @@ package azure
 
 import (
 	"giantswarm/dex-operator/pkg/idp/provider"
-	"giantswarm/dex-operator/pkg/key"
 	"strconv"
 	"testing"
 	"time"
@@ -36,7 +35,7 @@ func TestGetRequestBody(t *testing.T) {
 			if uri[0] != tc.config.RedirectURI {
 				t.Fatalf("Expected %s, got %v", tc.config.RedirectURI, uri[0])
 			}
-			s := GetSecretCreateRequestBody(tc.config.Name, key.SecretValidityMonths)
+			s := GetSecretCreateRequestBody(tc.config)
 			secretName := s.GetPasswordCredential().GetDisplayName()
 			if *secretName != tc.config.Name {
 				t.Fatalf("Expected %s, got %v", tc.config.Name, *secretName)