+++ pkg/idp/provider/github/github.go
@@ -41,13 +41,13 @@ type Github struct {
 	secret       string
 }
 
-type config struct {
-	organization string
-	team         string
-	appID        int64
-	privateKey   []byte
-	clientID     string
-	clientSecret string
+type Config struct {
+	Organization string
+	Team         string
+	AppID        int64
+	PrivateKey   []byte
+	ClientID     string
+	ClientSecret string
 }
 
 func New(p provider.ProviderCredential, log *logr.Logger) (*Github, error) {
@@ -59,7 +59,7 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Github, error) {
 	}
 
 	// get the client
-	itr, err := ghinstallation.NewAppsTransport(http.DefaultTransport, c.appID, c.privateKey)
+	itr, err := ghinstallation.NewAppsTransport(http.DefaultTransport, c.AppID, c.PrivateKey)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -71,37 +71,37 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Github, error) {
 		Type:         ProviderConnectorType,
 		Client:       client,
 		Owner:        p.Owner,
-		Organization: c.organization,
-		Team:         c.team,
-		id:           c.clientID,
-		secret:       c.clientSecret,
+		Organization: c.Organization,
+		Team:         c.Team,
+		id:           c.ClientID,
+		secret:       c.ClientSecret,
 	}, nil
 }
 
-func newGithubConfig(p provider.ProviderCredential, log *logr.Logger) (config, error) {
+func newGithubConfig(p provider.ProviderCredential, log *logr.Logger) (Config, error) {
 	if log == nil {
-		return config{}, microerror.Maskf(invalidConfigError, "Logger must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Logger must not be empty.")
 	}
 	if p.Name == "" {
-		return config{}, microerror.Maskf(invalidConfigError, "Credential name must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Credential name must not be empty.")
 	}
 	if p.Owner == "" {
-		return config{}, microerror.Maskf(invalidConfigError, "Credential owner must not be empty.")
+		return Config{}, microerror.Maskf(invalidConfigError, "Credential owner must not be empty.")
 	}
 
 	var organization, team, clientSecret, clientID string
 	{
 		if organization = p.Credentials[OrganizationKey]; organization == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", OrganizationKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", OrganizationKey)
 		}
 		if team = p.Credentials[TeamKey]; team == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", TeamKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", TeamKey)
 		}
 		if clientSecret = p.Credentials[ClientSecretKey]; clientSecret == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientSecretKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientSecretKey)
 		}
 		if clientID = p.Credentials[ClientIDKey]; clientID == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientIDKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", ClientIDKey)
 		}
 	}
 
@@ -109,10 +109,10 @@ func newGithubConfig(p provider.ProviderCredential, log *logr.Logger) (config, e
 	{
 		var err error
 		if appIDvalue := p.Credentials[AppIDKey]; appIDvalue == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", AppIDKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", AppIDKey)
 		} else {
 			if appID, err = strconv.Atoi(appIDvalue); err != nil {
-				return config{}, microerror.Maskf(invalidConfigError, "%s is not a valid value for %s: %v", appIDvalue, AppIDKey, err)
+				return Config{}, microerror.Maskf(invalidConfigError, "%s is not a valid value for %s: %v", appIDvalue, AppIDKey, err)
 			}
 		}
 	}
@@ -120,24 +120,28 @@ func newGithubConfig(p provider.ProviderCredential, log *logr.Logger) (config, e
 	var privateKey []byte
 	{
 		if privateKeyValue := p.Credentials[PrivateKeyKey]; privateKeyValue == "" {
-			return config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", PrivateKeyKey)
+			return Config{}, microerror.Maskf(invalidConfigError, "%s must not be empty.", PrivateKeyKey)
 		} else {
 			privateKey = []byte(privateKeyValue)
 		}
 	}
 
-	return config{
-		organization: organization,
-		team:         team,
-		appID:        int64(appID),
-		privateKey:   privateKey,
-		clientSecret: clientSecret,
-		clientID:     clientID,
+	return Config{
+		Organization: organization,
+		Team:         team,
+		AppID:        int64(appID),
+		PrivateKey:   privateKey,
+		ClientSecret: clientSecret,
+		ClientID:     clientID,
 	}, nil
 }
 
 func (g *Github) GetName() string {
 	return g.Name
+}
+
+func (g *Github) GetProviderName() string {
+	return ProviderName
 }
 
 func (g *Github) GetType() string {
@@ -249,4 +253,19 @@ func permissionsUpdateNeeded(app *githubclient.App) bool {
 func callbackURIPresent(app *githubclient.App, config provider.AppConfig) bool {
 	//TODO: check if callback URL is present
 	return true
+}
+
+func (g *Github) GetCredentialsForAuthenticatedApp(config provider.AppConfig) (string, error) {
+	// TODO: manifest flow
+	c := Config{}
+	return fmt.Sprintf(`client-id: %s
+client-secret: %s
+organization: %s
+team: %s
+app-id: %v
+private-key: %s`, c.ClientID, c.ClientSecret, c.Organization, c.Team, c.AppID, c.PrivateKey), nil
+}
+func (g *Github) CleanCredentialsForAuthenticatedApp(config provider.AppConfig) error {
+	// TODO open for deletion?
+	return nil
 }