+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `setup` package with functionality to create dex-operator credentials
+- Add credential creation, update and cleanup functionality to azure provider
+- Add github provider
+
 ## [0.2.0] - 2023-02-27
 
 ### Added
 
-- Add github provider
 - Add use of runtime/default seccomp profile.
 
 ### Changed